pyproject.toml
[project]
name = "long-context"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.99.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "yt-dlp>=2025.8.11",
]
